// src/pages/Manager/CourseRecord/data.d.ts
/*
 * @,@Author: ,: your name
 * @,@Date: ,: 2021-05-27 08:49:34
 * @,@LastEditTime: ,: 2021-05-27 09:54:16
 * @,@LastEditors: ,: Please set LastEditors
 * @,@Description: ,: In User Settings Edit
 * @,@FilePath: ,: \afterClass\src\pages\Manager\Statistics\AfterSchoolCourse\data.d.ts
 */

export type TableItem = {
  title: string,
  dataIndex: string,
  valueType?: string,
  width?: number,
  key?: string,
  align: string,
};